news-aggregator/requirements.txt
feedparser
google-cloud-aiplatform
python-dotenv
pyyaml
slack-sdk